arreglos de user

+++ be/apps/core/views/user.py
@@ -54,36 +54,40 @@ class UserView(View):
             except json.JSONDecodeError:
                 return JsonResponse({"error": "Invalid JSON"}, status=400)
         elif not request.user.is_authenticated:
+            print("%d\n", request.user)
             return JsonResponse({"error": "You do not have permission to modify this user's data."}, status=401)
         else:
             return JsonResponse({"error": "You do not have permission to modify this user's data."}, status=403)
-#hay que añadir la comprobacion de que el user que quiere eliminar es el mismo user
-    def delete(self, _, user_id):
+
+    def delete(self, request: HttpRequest, user_id):
         user = get_object_or_404(User, id=user_id)
-        anon_user, _ = User.objects.get_or_create(username="deleted_user")
-        random_numb = time.time() + random.uniform(0.1, 1.0)
-        user.password = ""
-        user.avatar = ""#esto hay que buscar
-        user.email = f"deleted_user_{random_numb}@anon.com"
-        user.status = False
-        user.deleted_user = True
-        #delete user friends
-        friends = Friends.objects.filter(user_id=user_id)
-        friends.delete()
-        #delete user from friend lists
-        for friend in Friends.objects.filter(friend_id=user_id):
-            friend.friend_id = anon_user
-        #delete user history
-        history = History.objects.filter(user_id=user_id)
-        history.delete()
-        #delete the user from other users' history lists
-        for history in History.objects.filter(opponent=user_id):
-            history.opponent = anon_user
-        #delete the user from tournaments
-        for tournament in Tournaments.objects.filter(players=user):
-            tournament.players.remove(user)
-            tournament.players.add(anon_user)
-        user.username = "user_" + str(random_numb)
-        user.save()
-        return JsonResponse({"message": "Your account and all associated data have been permanently deleted."}, status=204)
+        if request.user.is_authenticated and request.user.id == user_id:
+            anon_user, _ = User.objects.get_or_create(username="deleted_user")
+            random_numb = time.time() + random.uniform(0.1, 1.0)
+            user.password = ""
+            user.avatar = ""#esto hay que buscar
+            user.email = f"deleted_user_{random_numb}@anon.com"
+            user.status = False
+            user.deleted_user = True
+            #delete user friends
+            friends = Friends.objects.filter(user_id=user_id)
+            friends.delete()
+            #delete user from friend lists
+            for friend in Friends.objects.filter(friend_id=user_id):
+                friend.friend_id = anon_user
+            #delete user history
+            history = History.objects.filter(user_id=user_id)
+            history.delete()
+            #delete the user from other users' history lists
+            for history in History.objects.filter(opponent=user_id):
+                history.opponent = anon_user
+            #delete the user from tournaments
+            for tournament in Tournaments.objects.filter(players=user):
+                tournament.players.remove(user)
+                tournament.players.add(anon_user)
+            user.username = "user_" + str(random_numb)
+            user.save()
+            return JsonResponse({"message": "Your account and all associated data have been permanently deleted."}, status=204)
+        else:
+            return JsonResponse({"error": "You do not have permission to delete this user's data."}, status=403)
